slight cleanup

+++ bbarchivist/barutils.py
@@ -497,10 +497,9 @@ def remove_signed_files(a_folder):
     """
     files = [os.path.join(os.path.abspath(a_folder), file) for file in os.listdir(a_folder)]
     for file in files:
-        if file.endswith(".signed"):
-            if os.path.exists(file):
-                print("REMOVING: " + file)
-                os.remove(os.path.abspath(file))
+        if file.endswith(".signed") and os.path.exists(file):
+            print("REMOVING: " + file)
+            os.remove(os.path.abspath(file))
 
 
 def remove_unpacked_loaders(osdir, raddir, radios):
